chore: move drag-handle logic to selection-manager
