+++ packages/blocks/src/page-block/edgeless/selection-manager.ts
@@ -1,13 +1,20 @@
+import type {
+  BlockComponentElement,
+  MouseMode,
+  SelectionEvent,
+  TopLevelBlockModel,
+} from '@blocksuite/blocks/std';
+import {
+  getClosestBlockElementByPoint,
+  getModelByBlockElement,
+  getSelectedStateRectByBlockElement,
+  initMouseEventHandlers,
+  noop,
+  Point,
+} from '@blocksuite/blocks/std';
 import type { PhasorElement } from '@blocksuite/phasor';
 import type { Page } from '@blocksuite/store';
 
-import {
-  initMouseEventHandlers,
-  type MouseMode,
-  noop,
-  type SelectionEvent,
-  type TopLevelBlockModel,
-} from '../../__internal__/index.js';
 import { updateLocalSelectionRange } from '../default/selection-manager/utils.js';
 import type { EdgelessPageBlockComponent } from './edgeless-page-block.js';
 import { BrushModeController } from './mode-controllers/brush-mode.js';
@@ -161,6 +168,30 @@ export class EdgelessSelectionManager {
   };
 
   private _onContainerMouseMove = (e: SelectionEvent) => {
+    const point = new Point(e.raw.clientX, e.raw.clientY);
+    point.x += 26 - 1;
+    let hoverEditingState = null;
+    let element = getClosestBlockElementByPoint(
+      point
+      // this._container.innerRect
+    );
+    if (element) {
+      if (element.tagName === 'AFFINE-EDGELESS-PAGE') {
+        element = null;
+      }
+      if (element) {
+        hoverEditingState = {
+          element: element as BlockComponentElement,
+          model: getModelByBlockElement(element),
+          rect: getSelectedStateRectByBlockElement(element),
+        };
+      }
+    }
+    this._container.components.dragHandle?.onContainerMouseMove(
+      e,
+      hoverEditingState
+    );
+
     this._updateLastMousePos(e);
     this._container.slots.hoverUpdated.emit();
     return this._controllers[this.mouseMode.type].onContainerMouseMove(e);
